requirements.txt
numpy
opencv-python
pyzbar
rpi.gpio